refactor(ui): move types to common folder
