+++ src/framework/ui/overflowMenu/overflowMenu.component.tsx
@@ -18,7 +18,7 @@ import {
   Popover as PopoverComponent,
   Props as PopoverProps,
 } from '../popover/popover.component';
-import { Omit } from '../service/type';
+import { Omit } from '../common/type';
 
 type MenuItemElement = React.ReactElement<OverflowMenuItemProps>;
